rename github instances to git
