+++ frontend/__tests__/components/features/github/github-repo-selector.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 import { renderWithProviders } from "test-utils";
-import { GitRepositorySelector } from "#/components/features/github/github-repo-selector";
+import { GitRepositorySelector } from "#/components/features/github/git-repo-selector";
 import OpenHands from "#/api/open-hands";
 import { Provider } from "#/types/settings";
