+++ frontend/src/api/open-hands-axios.ts
@@ -7,18 +7,8 @@ export const setAuthTokenHeader = (token: string) => {
   openHands.defaults.headers.common.Authorization = `Bearer ${token}`;
 };
 
-export const setGitHubTokenHeader = (token: string) => {
-  openHands.defaults.headers.common["X-GitHub-Token"] = token;
-};
-
 export const removeAuthTokenHeader = () => {
   if (openHands.defaults.headers.common.Authorization) {
     delete openHands.defaults.headers.common.Authorization;
-  }
-};
-
-export const removeGitHubTokenHeader = () => {
-  if (openHands.defaults.headers.common["X-GitHub-Token"]) {
-    delete openHands.defaults.headers.common["X-GitHub-Token"];
   }
 };
