+++ frontend/src/routes/_oh._index/route.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { useDispatch } from "react-redux";
 import posthog from "posthog-js";
 import { setReplayJson } from "#/state/initial-query-slice";
-import { useGitHubUser } from "#/hooks/query/use-github-user";
+import { useGitUser } from "#/hooks/query/use-github-user";
 import { useGitHubAuthUrl } from "#/hooks/use-github-auth-url";
 import { useConfig } from "#/hooks/query/use-config";
-import { GitRepositoriesSuggestionBox } from "#/components/features/github/github-repositories-suggestion-box";
+import { GitRepositoriesSuggestionBox } from "#/components/features/github/git-repositories-suggestion-box";
 import { ReplaySuggestionBox } from "../../components/features/suggestions/replay-suggestion-box";
 import { CodeNotInGitHubLink } from "#/components/features/github/code-not-in-github-link";
 import { HeroHeading } from "#/components/shared/hero-heading";
@@ -18,7 +18,7 @@ function Home() {
   const formRef = React.useRef<HTMLFormElement>(null);
 
   const { data: config } = useConfig();
-  const { data: user } = useGitHubUser();
+  const { data: user } = useGitUser();
 
   const gitHubAuthUrl = useGitHubAuthUrl({
     appMode: config?.APP_MODE || null,
